requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
filelock==3.12.0
Flask
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
packaging==23.1
platformdirs==3.5.1
psycopg2==2.9.6
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
virtualenv==20.23.0
WTForms==3.0.1
